+++ config.php
@@ -26,7 +26,7 @@ function config(RectorConfig $rectorConfig): void
 function laravel(RectorConfig $rectorConfig): void
 {
     config($rectorConfig);
-    $rectorConfig->rule(\MLL\RectorConfig\Rector\PrimaryKeyColumnToIdAliasRector::class);
+    $rectorConfig->rule(\MLL\RectorConfig\Rector\PrimaryKeyColumnToIDAliasRector::class);
     $rectorConfig->ruleWithConfiguration(\Rector\Transform\Rector\FuncCall\FuncCallToNewRector::class, [
         'collect' => \Illuminate\Support\Collection::class,
     ]);