+++ tests/Rector/PrimaryKeyColumnToIDAliasRector/Source/ReadOnlyLegacyAccessorExamination.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace MLL\RectorConfig\Tests\Rector\PrimaryKeyColumnToIdAliasRector\Source;
+namespace MLL\RectorConfig\Tests\Rector\PrimaryKeyColumnToIDAliasRector\Source;
 
 use Illuminate\Database\Eloquent\Model;
 
-class LegacyAccessorExamination extends Model
+class ReadOnlyLegacyAccessorExamination extends Model
 {
     /** @var string */
     protected $primaryKey = 'exam_no';
@@ -12,10 +12,5 @@ class LegacyAccessorExamination extends Model
     public function getIdAttribute(): ?string
     {
         return $this->attributes[$this->primaryKey] ?? null;
-    }
-
-    public function setIdAttribute(?string $id): void
-    {
-        $this->attributes[$this->primaryKey] = $id;
     }
 }