+++ tests/Rector/PrimaryKeyColumnToIDAliasRector/config/configured_rule.php
@@ -1,8 +1,8 @@
 <?php declare(strict_types=1);
 
-use MLL\RectorConfig\Rector\PrimaryKeyColumnToIdAliasRector;
+use MLL\RectorConfig\Rector\PrimaryKeyColumnToIDAliasRector;
 use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(PrimaryKeyColumnToIdAliasRector::class);
+    $rectorConfig->rule(PrimaryKeyColumnToIDAliasRector::class);
 };